+++ assets/bankregistration.sql
