--- bankWebsite/contact.html
@@ -1,126 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Contact Us</title>
-  <link rel="stylesheet" type="text/css" href="contact.css">
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,200;0,400;1,200&display=swap"
-    rel="stylesheet">
-</head>
-
-<body>
-  <header>
-    <a href ="homepage.html" class="logo">NOIR CAPITAL BANK</a>
-
-    <ul class="navlist">
-      <li><a href="homepage.html">Home</a></li>
-      <li><a href="login.html">Sign In</a></li>
-      <li><a href="contact.html">Contact</a></li>
-      <li><a href="about.html">About</a></li>
-      <li><a href="atm.html">ATM</a></li>
-    </ul>
-    <div class="bx bx-menu" id="menu-icon"></div>
-  </header>
-
-  <br>
-  <br>
-  <br>
-
-
-  <div class="container">
-    <h1>Contact Us</h1>
-    <p>Feel free to contact us through our email and/or number, and a message of your inquiry.</p>
-    <!--<form  method="post"> 
-      <br>
-      <label for="name">Name:</label>
-      <input type="text" id="name" name="name" required>
-      <br>
-      <label for="email">Email:</label>
-      <input type="email" id="email" name="email" required>
-       <br>
-      <label for="message">Message:</label>
-      <textarea id="message" name="message" required></textarea>
-
-      <input type="submit" value="Submit">
-  </div>
-    -->
-
-    <div class="contact-box">
-      <div class="contact-left">
-        <h3>Ask a Question!</h3>
-        <form action="contact.php" method="post">
-          <div class="input-row">
-            <div class="input-group">
-              <label>Name </label>
-              <input type="text" placeholder="Name">
-            </div>
-
-
-            <div class="input-group">
-              <label>Phone </label>
-              <input type="text" placeholder="415-0000-0000">
-            </div>
-          </div>
-
-
-          <div class="input-row">
-            <div class="input-group">
-              <label>Email </label>
-              <input type="email" placeholder="sjsu@sjsu.edu.com">
-            </div>
-
-            <div class="input-group">
-              <label>Message</label>
-              <input type="text" placeholder="Message">
-            </div>
-          </div>
-
-          <label>Inquiry</label>
-          <textarea rows="6" placeholder="Input Message"> </textarea>
-
-          <button type="submit">Submit</button>
-
-        </form>
-
-      </div>
-
-
-      <div class="contact-right">
-        <h3>Contact Information</h3>
-        <table>
-          <tr>
-            <td>Email: </td>
-            <td>sjsu@sjsu.edu</td>
-
-          </tr>
-
-          <tr>
-            <td>Number: </td>
-
-            <td>415-0000-000</td>
-            <br>
-          </tr>
-
-          <tr>
-            <td>Address: </td>
-            <td>1 Washington Sq, San Jose, CA 95192</td>
-            <br>
-          </tr>
-
-        </table>
-
-      </div>
-
-
-    </div>
-
-    <footer class="footer">
-      &copy; 2023 NOIR CAPITAL BANK. All rights reserved.
-    </footer>
-    <script src="script.js"></script>
-</body>
-
-</html>
