--- bankWebsite/passwordCheck.js
@@ -1,12 +1,0 @@
-var password = document.getElementById("password"), confirm_password = document.getElementById("cpassword");
-
-function validatePassword(){
-  if(password.value != confirm_password.value) {
-    confirm_password.setCustomValidity("Passwords Don't Match");
-  } else {
-    confirm_password.setCustomValidity('');
-  }
-}
-
-password.onchange = validatePassword;
-confirm_password.onkeyup = validatePassword;
