+++ css/contact.css
@@ -1,0 +1,194 @@
+.container {
+    margin: 50px auto;
+    width: 80%;
+    padding: 10px;
+  }
+  
+  * {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+    list-style: none;
+    text-decoration: none;
+  }
+  
+  header {
+    position: fixed;
+    right: 0;
+    top: 0;
+    z-index: 1000;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 33px 9%;
+    background: #808080;
+  }
+  
+  .logo {
+    font-size: 30px;
+    font-weight: 700;
+    color: white;
+  }
+  
+  .navlist {
+    display: flex;
+  }
+  
+  .navlist a {
+    color: white;
+    margin-left: 60px;
+    font-size: 15px;
+    font-weight: 600;
+    border-bottom: 2px solid transparent;
+    transition: all .55s ease;
+  }
+  
+  navlist a:hover {
+    border-bottom: 2px solid white;
+  }
+  
+  #menu.icon {
+    color: white;
+    font-size: 30px;
+    z-index: 10001;
+    cursor: pointer;
+    display: none;
+  }
+  
+  footer {
+    background: #555;
+    color: #fff;
+    padding: 1em;
+    text-align: center;
+    bottom: 0;
+    width: 100%;
+  }
+  
+  body {
+    background: #f2f2f2;
+    font-family: 'Poppins', sans-serif;
+    margin: 0;
+    padding: 0;
+  
+  }
+  
+  .contact-box {
+    background: white;
+    display: flex;
+  }
+  
+  .contact-left {
+    flex-basis: 60%;
+    padding: 40px 60px;
+  }
+  
+  .contact-right {
+    flex-basis: 40%;
+    padding: 40px 40px;
+    background: #008fff;
+  }
+  
+  h1 {
+    margin-bottom: 10px;
+  }
+  
+  .container p {
+    margin-bottom: 40px;
+  }
+  
+  .input-row {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 20px;
+  }
+  
+  .input-row .input-group {
+    flex-basis: 45%;
+  }
+  
+  input {
+    width: 100%;
+    border: none;
+    border-bottom: 1px solid #ccc;
+    outline: none;
+    padding-bottom: 6px;
+  }
+  
+  textarea {
+    width: 100%;
+    border: 1px solid #ccc;
+    outline: none;
+    padding: 10px;
+    box-sizing: border-box;
+  }
+  
+  label {
+    margin-bottom: 6px;
+    display: block;
+    color: #1c00b5
+  }
+  
+  button {
+    background: #1c00b5;
+    width: 100px;
+    border: none;
+    outline: none;
+    color: #fff;
+    height: 35px;
+    border-radius: 30px;
+    margin-top: 20px;
+    box-shadow: 0px 5px 15px 0px rgba(28, 0, 181, 0.3);
+  }
+  
+  .contact-left h3 {
+    color: #1c00b5;
+    font-weight: 600;
+    margin-bottom: 10px;
+  }
+  
+  .contact-right h3 {
+    font-weight: 600;
+    margin-bottom: 10px;
+  }
+
+  .popup{
+    width: 450px;
+    background: #d4e6f3;
+    border-radius: 6px;
+    position: absolute;
+    top: 0%;
+    left: 50%;
+    transform: translate(-50%,-50%) scale(0.1);
+    text-align: center;
+    padding: 0 30px 30px;
+    color: #333;
+    visibility: hidden;
+    transition: transform 0.4s, top 0.4s;
+    border: 6px solid transparent; 
+
+     /* Create a linear gradient border */
+    border-image: linear-gradient(to right,#8fb0c7, #c4daeb, #ceced1,#c4daeb, #bce2eb);
+    border-image-slice: 1;
+    border-image-width: 6px;
+  }
+
+  .popup img{
+    width: 200px;
+    margin-top: -50px;
+    border-radius: 50%
+    
+  };
+
+  .popup h2{
+    font-size: 38px;
+    font-weight: 500;
+    margin: 30px 0 10px;
+  }
+
+  .open-popup{
+    visibility: visible;
+    top: 50%;
+    transform: translate(-50%,-50%)scale(1);
+  }
