+++ webPage/about.html
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>About</title>
+
+    <link rel="stylesheet" type="text/css" href="../css/about.css" />
+
+    <link
+      rel="stylesheet"
+      href="https://unpkg.com/boxicons@latest/css/boxicons.min.css"
+    />
+
+    <link
+      href="https://cdn.jsdelivr.net/npm/remixicon@2.5.0/fonts/remixicon.css"
+      rel="stylesheet"
+    />
+
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400&display=swap"
+      rel="stylesheet"
+    />
+  </head>
+
+  <body>
+    <header>
+      <a href="../webPage/homepage.html" class="logo">NOIR CAPITAL BANK</a>
+
+      <ul class="navlist">
+        <li><a href="../webPage/homepage.html">Home</a></li>
+        <li><a href="../webPage/login.php">Sign In</a></li>
+        <li><a href="../webPage/contact.html">Contact</a></li>
+        <li><a href="../webPage/about.html">About</a></li>
+        <li><a href="../webPage/atm.html">ATM</a></li>
+      </ul>
+      <div class="bx bx-menu" id="menu-icon"></div>
+    </header>
+
+    <section class="bank">
+      <div class="bank-text">
+        <h4>About Us</h4>
+        <h1>What do we do?</h1>
+        <p>
+          Welcome to Noir Capital Bank, your trusted financial partner. We're
+          committed to delivering secure, convenient, and innovative banking
+          solutions. Our mission is to empower you to achieve your financial
+          goals. We recognize the importance of security, convenience,
+          innovation, and transparency. Join us today!
+        </p>
+        <a href="../webPage/registration.html">Register</a>
+        <a href="#" class="ctaa"></a>
+      </div>
+      <div class="text-body"></div>
+    </section>
+    <div class="icons"></div>
+
+    <script src="https://unpkg.com/scrollreveal"></script>
+
+    <script src="../js/home.js"></script>
+    <footer>&copy; 2023 NOIR CAPITAL BANK. All rights reserved.</footer>
+  </body>
+</html>
