+++ webPage/atm.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Accounts - ATM</title>
+
+    <link rel="stylesheet" type="text/css" href="../css/home.css" />
+
+    <link
+      rel="stylesheet"
+      href="https://unpkg.com/boxicons@latest/css/boxicons.min.css"
+    />
+
+    <link
+      href="https://cdn.jsdelivr.net/npm/remixicon@2.5.0/fonts/remixicon.css"
+      rel="stylesheet"
+    />
+
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400&display=swap"
+      rel="stylesheet"
+    />
+  </head>
+
+  <body>
+    <header>
+      <a href="../webPage/homepage.html" class="logo">NOIR CAPITAL BANK</a>
+
+      <ul class="navlist">
+        <li><a href="../webPage/homepage.html">Home</a></li>
+        <li><a href="../webPage/login.php">Sign In</a></li>
+        <li><a href="../webPage/contact1.php">Contact</a></li>
+        <li><a href="../webPage/about.html">About</a></li>
+        <li><a href="../webPage/atm.html">ATM</a></li>
+      </ul>
+      <div class="bx bx-menu" id="menu-icon"></div>
+    </header>
+    <section class="bank">
+      <div class="atm">
+        <h2>Virtual ATM</h2>
+        <div class="atm-screen">
+          <div class="input-field">
+            <label for="cardNumber">Card Number:</label>
+            <input
+              type="text"
+              id="cardNumber"
+              placeholder="Enter your card number"
+            />
+          </div>
+          <div class="input-field">
+            <label for="pin">PIN:</label>
+            <input type="password" id="pin" placeholder="Enter your PIN" />
+          </div>
+          <button onclick="withdraw()">Submit</button>
+        </div>
+        <br />
+        <div class="buttons">
+          <div class="row">
+            <button id="back">Back</button>
+          </div>
+          <div class="row">
+            <button id="but">WITHDRAW</button>
+            <button id="but">DEPOSIT</button>
+            <button id="but">BAL INQUIRY</button>
+            <button id="but">CHANGE PIN</button>
+            <button id="but">DETAILS</button>
+          </div>
+          <div class="row">
+            <button id="cancel">Cancel</button>
+          </div>
+        </div>
+      </div>
+    </section>
+    <div class="icons"></div>
+
+    <script src="https://unpkg.com/scrollreveal"></script>
+
+    <script src="../js/home.js"></script>
+    <footer>&copy; 2023 NOIR CAPITAL BANK. All rights reserved.</footer>
+  </body>
+</html>
