+++ webPage/homepage.html
@@ -1,0 +1,67 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>NOIR - Home</title>
+
+    <link rel="stylesheet" type="text/css" href="../css/home.css" />
+
+    <link
+      rel="stylesheet"
+      href="https://unpkg.com/boxicons@latest/css/boxicons.min.css"
+    />
+
+    <link
+      href="https://cdn.jsdelivr.net/npm/remixicon@2.5.0/fonts/remixicon.css"
+      rel="stylesheet"
+    />
+
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400&display=swap"
+      rel="stylesheet"
+    />
+  </head>
+
+  <body>
+    <header>
+      <a href="../webPage/homepage.html" class="logo">NOIR CAPITAL BANK</a>
+
+      <ul class="navlist">
+        <li><a href="../webPage/homepage.html">Home</a></li>
+        <li><a href="../webPage/login.php">Sign In</a></li>
+        <li><a href="../webPage/contact1.php">Contact</a></li>
+        <li><a href="../webPage/about.html">About</a></li>
+        <li><a href="../webPage/atm.html">ATM</a></li>
+      </ul>
+      <div class="bx bx-menu" id="menu-icon"></div>
+    </header>
+
+    <section class="bank">
+      <div class="bank-text">
+        <h5>"Bank Anywhere, Anytime: Your Money, Your Rules."</h5>
+        <h4>Welcome!</h4>
+        <h1>The Future Of Banking Is Here.</h1>
+        <p>
+          Begin your online financial journey today. Your Noir Capital Bank
+          Account will be all you need for spending, saving, depositing,
+          withdrawing, and transfers. Join us today!
+        </p>
+        <a href="#" onclick="openPopOutWindow()">Register</a>
+        <a href="../webPage/login.php">Login</a>
+      </div>
+
+      <div class="bank-img">
+        <img src="../assets/home.png" />
+      </div>
+    </section>
+    <div class="icons"></div>
+
+    <script src="https://unpkg.com/scrollreveal"></script>
+
+    <script src="../js/home.js"></script>
+    <footer>&copy; 2023 NOIR CAPITAL BANK. All rights reserved.</footer>
+  </body>
+</html>
