+++ webPage/registration.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>NOIR CAPITAL BANK -- Register</title>
-    <link rel="stylesheet" type="text/css" href="home.css" />
+    <link rel="stylesheet" type="text/css" href="../css/home.css" />
 
     <link
       rel="stylesheet"
@@ -139,7 +139,7 @@
       <!-- this is the animation at the top -->
       <div id="divChange"></div>
       <div id="container">
-        <form action="/registrationprocess.php" method="post">
+        <form action="../processingPages/registrationprocess.php" method="post">
           <label id="titleLabel">User Registration</label>
           <br />
           <br />
@@ -180,7 +180,7 @@
             required
           />
 
-          <script src="passwordCheck.js"></script>
+          <script src="../js/passwordCheck.js"></script>
 
           <br />
           <label id="textLabel">Phone Number</label>
